Update tron.bat
I added lines 1046-1058.  Check licenses and test before pushing.  I have not tested this bat.  I have tested the ahk scripts with MBAM 2.0.4.1028 on a Windows 8.1 system.

http://ahkscript.org/download/ahk-u32.zip = AutoHotKey32.exe
http://ahkscript.org/download/ahk-u64.zip = AutoHotKey64.exe
After unzipping, they would need to be renamed to such and placed in the mbam folder to work.  I am not sure if 64 matters in this instance.  They may both work with 32.